are you sure functionality

+++ public/pages/profile.php
@@ -56,7 +56,7 @@ $uid = $_SESSION['uid'] ?? null;
                 <?php
                 if($utype === 1){
                     echo "<div id=\"icon-buttons\">
-                    <a href=\"\" class=\"link-button\" id=\"delete-profile-button\"><i
+                    <a href=\"\" class=\"link-button\" id=\"delete-profile-button\" onclick=\"return confirm('Are you sure?')\"><i
                             class=\"fa-regular fa-trash-can\"></i></a>
                 </div>";
                 
@@ -82,7 +82,7 @@ $uid = $_SESSION['uid'] ?? null;
                     echo " </article>";
                     if (isset($postImg)) { echo "<img src=\"$postImg\">";}
                     if (($utype === 0 && $uid == $userId) || $utype === 1) {
-                        echo "<div id=\"icon-buttons\"> <a href=\"../scripts/delete_my_posts.php?postId=$pid\" class=\"link-button\" id=\"delete-post-button\"><i class=\"fa-regular fa-trash-can\"></i></a></div>";
+                        echo "<div id=\"icon-buttons\"> <a href=\"../scripts/delete_my_posts.php?postId=$pid\" class=\"link-button\" id=\"delete-post-button\" onclick=\"return confirm('Are you sure?')\"><i class=\"fa-regular fa-trash-can\"></i></a></div>";
                     }
                     echo "</div>";
                     while ($prstmt->fetch()) {
@@ -94,7 +94,7 @@ $uid = $_SESSION['uid'] ?? null;
                         echo " </article>";
                         if (isset($postImg)) { echo "<img src=\"$postImg\">";}
                         if (($utype === 0 && $uid == $userId) || $utype === 1) {
-                            echo "<div id=\"icon-buttons\"> <a href=\"../scripts/delete_my_posts.php?postId=$pid\" class=\"link-button\" id=\"delete-post-button\"><i class=\"fa-regular fa-trash-can\"></i></a></div>";
+                            echo "<div id=\"icon-buttons\"> <a href=\"../scripts/delete_my_posts.php?postId=$pid\" class=\"link-button\" id=\"delete-post-button\" onclick=\"return confirm('Are you sure?')\"><i class=\"fa-regular fa-trash-can\"></i></a></div>";
                         }
                         echo "</div>";
                     }
